Configures auto-exporter to publish shaded and non-shaded variants (#235)
* Configures auto-exporter to publish shaded and non-shaded variants

* Remove shadowed=true from gradle properties

Since auto-exporter has to be published as both - shaded and unshaded
variant, the only required changes is to remove the shadowed property
from gradle which prevents publishing shaded variants.

The gradle shadow plugin automatically publishes a shaded artifact when
used with maven publish plugin. The sources and javadoc for shaded JAR
are not published unless explicitly mentioned.