Make project compatible with Java 17
Mockito 3.5.x is incompatible when building with Java 17, therefore an
update to a more recent version of Mockito is required. There have been
breaking changes in Mockito's releases which are difficult to point out.
Some tests required code changes to make them compatible with these
changes.
Only breaking change that affected this codebase was the handling of
static mocks.